ApplyMasksJob: support for larger tileshapes
Previously, tiles were limited to a maximum of one row of frames. Now
they can be as big as a whole partition. In combination with mmapped
datasets this allows for a nice speedup: L3 blocking is then done by the
BLAS implementation, which is native code and obviously quite well optimized.