ApplyMasksJob: support for more tile shapes
Before, the only tileshapes that were really supported were
of the kind (1, x, py, px) or (y, x, py, px), where tiles of (1, x, py, px)
could not cross rows, that is, they had to be rectangles in the scan
dimension. Now, we support tileshapes that cross rows, that is,
something like (1, x, py, px) where x is larger than the scan width,
or where x does not evenly divide the scan width.

We do this by reshaping the result array to a general (num_masks, num_frames)
and copying into that.