Select peaks above background from diffraction pattern
This notebook uses numpy to generate advanced masks to extract peaks
in a diffraction pattern. It uses the Python API to apply these masks
to a specific dataset

Each peak is surrounded by a reference region with negative weight
so that only the difference between peak and reference is recorded.
The result is a crisp dark field image
that integrates over many peaks for a better signal-to-noise ratio and
is insensitive to diffuse background scattering.