Notebook for Python prototype
This notebook encodes and decodes one block of little endian uint12
with an algorithm that should be suitable for vectorization,
i.e. it reads and writes aligned 32 bit uints and can process
several blocks in parallel with the same instructions.

Furthermore, it contains a function to calculate the mask and shift operations
to encode and decode little endian uints with n bits on a little endian CPU.