Modified and unified to work for big and little endian
Provided new Python code and updated README to cover all permutations of
big and little endian data with all bit lengths. The C++ code is now just a sample
of processing big endian uint12 on a little endian machine.

Note that the K2 data is LITTLE-Endian. This is covered in the manual
sample processing in the Jupyter notebook. The C++ code is not changed because
it still works as a sample. Pending work from @sk1p to build a proper C/C++
version that can process the actual K2 data.

TODO: Do we want to cover processing K2 data on a big endian machine efficiently?