Full encoder and decoder for little endian uint12 on a little endian machine
The code added to encode-decode.ipynb handles encoding and decoding
of arrays with arbitrary length with all index calculations
and treatment of dangling blocks. It was written to closely
mimick the way it would be written in C or C++ so that
it can serve as a prototype for a correct C/C++ implementation.