Work towards support for sparse matrices
This commit adds very basic support for sparse matrices

The modification of _make_mask_slicer() allows handling sparse matrices
since scipy.sparse matrices allow reshaping into matrices with (-1, 1),
not into vectors with (-1, ), and they support hstack() and vstack(),
but not stack(). The result of the function is the same.

The computation itself doesn't work yet for reasons that have to be
investigated. The test for sparse matrices allows to handle the situation
gracefully with a NotImplementedError instead of crashing.

The example currently triggers this error.