Working sparse matrix support
The example works and is significantly faster than the same operation
with dense matrices.

Mixing dense and sparse matrices is not tested yet. Possibly
a special sparse_mask_job as the only way to do sparse matrix
multiplication would be cleaner because that way we don't have to
handle mixtures?