Portable way to determine file owner
The pwd module is not available on Windows and uid and gid are 0 in
stat. Implement a platform-independent function in win_tweaks.py and
restructure the listing code since we need the full path to determine
an owner on Windows.

Make pypywin32 a platform-dependent dependency for Windows to support
listing the owner.