Include start of real cluster in tests
Add tests for both standard and subprocess method to start a local
cluster. The tests are only run if the LT_RUN_FUNCTIONAL
environment variable is set.

This is turned on in the Travis and Appveyor builds, in particular to
tun the new subprocess code in different environments.

To enable tests with a real cluster, we can't use unlinked tempfiles,
but we use a temporary directory from pytest where the files persist
for a whole session.

This will be useful for real data files as well.