Dimensionality refactoring (#195)
This change implements full back-end and API support for n-dimensional data sets.

Full credit goes to @sk1p !

The underlying assumption is that our data can be understood as C-order sequence of hypercuboids. The navigation dimension addresses a hypercuboid, and the signal dimension addresses data within a hypercuboid. The new Shape class manages this for the user.

The support includes now, as examples, line and vector scans, spectral and hyperspectral data, tilt series, time series, and non-rectangular detector data. Previously, only rectangular 2D scans with rectangular 2D detector frames were supported.

For many tasks, in particular mask application, the shape of the data is not relevant, as long as navigation and signal dimensions can be distinguished. Both navigation and signal dimension can then each be flattened so that the result is a 2D array. After mask application, the data is shaped back into the correct shape. As an example, a data set with shape (4, 5, 6, 8, 9), nav = (4, 5, 6), sig = (8, 9) accepts masks with shape (8, 9) and gives a result with shape (4, 5, 6).

The new system distinguishes between the Dataset.raw_shape of the underlying file format and the Dataset.shape which is the best representation towards the user. This particularly important for MIB data sets where individual files contain a specific number of frames regardless of the scan size. That way, the natural partitions of a MIB data set are not necessarily rectangular in the representation as scan. With this change, the back-end can work efficiently with the natural data layout of the Dataset, while the front-end and API presents the result in a way that is convenient for the user and for further processing.