[skip ci] Move to prototypes
Since this script uses a lot of the LiberTEM internal API, it is not
suitable as an "official" example. It can rather serve as a
prototype for the kind of functionality we'd like to offer in an
improved API.

Small optimization: Re-initialize the old memory with zeros
rather than build a new array. Faster and simpler.