Split up DataSet implementations
There is now a DataSetMeta class that should hold all metadata needed
from the dataset when running jobs, as well as *Reader classes where
needed, that implement the actual reading from files. Partition doesn't
get the whole DataSet anymore, only DataSetMeta.

This is an attempt to reduce the communication overhead of transferring
whole DataSet classes, which can be significant, especially if the
number of partitions is large.