Don't close over analysis instance in mask factories
When accessing 'self' in mask functions that are defined in analysis
subclasses, the whole analysis instance will be serialized, which
includes the DataSet. This can be costly, for example if the DataSet
includes a cache of some sorts. Observed with the MIB reader when
reading datasets with many small-ish files.

Also emit a warning if the mask factory closes over 'self'.

Refs #232