Optimized version using high-level array functions
Solvers, optimizers, dot product, array programming and such where possible.

According to profiling, it spends most of its time in the clustering algorithm.

The performance can be optimied for real world application by only clustering few coordinates, for
example from a sum image, and matching those candidates to many coordinates in the individual frames.

In that case it might make sense to look at unnecessary copies, memory layout and such for the matching.