Fit quality and weighted fit
The correlation code now calculates a fit quality function, and the grid matching code
uses that to eliminate really bad peaks and to assign weights.

Work in progress: The gridmatching implementation has grown rather messy and needs
a thorough cleanup.

Independent of that, the result in the jupyter notebook is looking very promising.