Example for performance effect of blocking numpy operations
The difference isa result of balancing overheads from
fine-grained loops with gains from working on data that fits
the L3 cache.

Fancy indexing with the boolean mask creates a copy. If the
copy fits the L3 cache, the subsequent calculation of std()
can use the data from the L3 cache, which is much faster.

This example shows how we should implement per-tile UDFs #310 in
order to reap the benefits of this, in particular for operations
that are limited by memory transfer. std() is such an operation.