Full file and metadata update on Zenodo.org
The script purges all files and adds new versions as given on the command line.

It uses a template in zenodo-upload.json and adds creators and contributors from respective
JSON files. These files already contain extra information that can be used to generate these parts
of the acknowledgments automatically in a next step. For that reason they are not part of the
template but separate files.

The description on Zenodo is taken from a semi-automatic HTML file made from README.rst. See also
https://github.com/zenodo/zenodo/issues/1736
regarding automatic descriptions.