UDF: prevent accidental dict changes
When writing a merge function, you have to take care not to set an item
in the dest dict, but instead write to the corrsponsing buffer. So this
code won't work:

    dest['pixelsum'] = src['pixelsum']

Instead, one should write:

    dest['pixelsum'][:] = src['pixelsum']

to explicitly write to the pixelsum buffer. By wrapping both the dest and
src arguments of the merge function in MappingProxyType, we prevent
accidentally setting a key in both of them.