Replace dedicated affine transformation code with equivalent that builds on existing methods
Replace the dedicated code for affine transformation refinement
with a mathematically equivalent version that builds on the
zero, a, b system.

It works with virtual base vectors such as zero=(cy, cx), a=(1, 0) and b=(0, 1)
and becomes completely equivalent to an affine transformation in that case.

It results in less code and re-uses existing methods.