Move the full frame refining to the prototypes folder
In tests, it didn't show a significantly better result in comparison to correlation-based
analysis of individual peaks, while being much, much slower.

Furthermore, this approach has a tendency to converge on "bad" maxima for very faint
and irregular CBED patterns.

As an exemplary application for UDFs and numba, and potentially as a basis to continue this work at a later time,
it could stay in the prototypes folder.