Disable sparse matrix for analyses
It turns out that the change in the sparse matrix backend
impacted the performance on single masks. Since deep mask stacks are
the most important application for sparse matrices and the
performance there was good, at least in the previous checks,
it is better to rely on dense masks for the sallow stacks for now.