Make AnalysisResultSet and AnalysisResult optionally lazy
The raw_results of RadialFourierAnalysis are more convenient for
scripting while the visualized and processed results are required for the GUI.

The visualization and AnalysisResults are calculated lazily with this change, which
allows working with many bins and orders for scripting without triggering visualization
overheads.

Furthermore, streamline mask backend decision logic further.