Adaptive rate for GUI update
Adapt the integration window for results to match the postprocessing and communication time.

This strongly reduces the update rate and increases the integration window if the network creates
backpressure or if calculating the result images is slow. That improves the overall performance and
user experience. We are not getting the results to the GUI anyway -- at least the merging is not
stalled and the next update will show more data when it finally comes.