Make cancellation explicit
This is done with a parameter cancel_id for both `run_job` and `cancel`
methods of the executor. For testing, this also introduces registering
dataset classes such that the web API can access them by name later. To
make this work, the dataset parameters are explicitly defined by
a schema and a conversion function ("json" -> python)