Create dask arrays from datasets; implement process_partitions()
This commit allows more control over the size of tiles, which allows
to create one tile for an entire partition. This is the basis for
creating dask arrays and implement process_partition for UDFs.

WIP: Test if this works properly remotely.
FIXME: Documentation, since this is a major integration feature for other packages like Hyperspy.