Test cases for ROI and postprocess() in combination with AUX data
These test cases are failing at the moment.

Observed behavior:

AUX data in form of a BufferWrapper seems to be transmitted in full to each worker, regardless
of partition or ROI. The AUX data is not masked corectly for ROI in process_frame(). The
AUX data is not masked at all in postprocess().

Expected behavior:
Ideally, only the part of the BufferWrapper that is relevant for each partition should
be transmitted to the worker, i.e. masked by partition and ROI. Otherwise the BufferWrapper should
at least be masked with the ROI when process_frame() etc is executed, and masked with partition and
ROI when postprocess() is executed.

Analysis:
Looking at the execution flow, the result buffers are allocated correctly masked for partition and ROI,
while the params are just passed through despite beging BufferWrapper objects as well. The included Jupyter
notebook helps to visualize the behavior.

In set_views_for_<...>, both result buffers and params are processed the same way despite having a different content shape.
This never triggers an error since the AUX data is always the same size or larger in
comparison to the result buffers.

Proposed fix:
Mask the result buffers by partition and ROI at an appropriate point, for example when generating
UDFTask objects. @sk1p, this is a bit above my pay grade. Could you please help with that?