Add a reset button to open dataset forms, fixes #433
The reset button switches the initial data for the form and the dataset
type to the detected parameters. It keeps the name intact, as that is
(not yet) detected from the filename (#246).

This also includes an update of all remaining nodejs dependencies.
Because of an updated typescript version, some small functions had
to be rewritten or re-typed.