Fix issues with cross-platform remote opening
It appears that the function used in executor.run_function() should be
defined in the module or class instead of being used directly in all cases
where it does platform-dependent things.

Just my guessing: This will make sure that the platform-dependent function
is not pickled and sent over, but that the function of the module on the
worker node is used instead.