Import color wheels from empyre instead of using a copy
Empyre has been released since the copy was created and we can use it
directly now.

LiberTEM contained a 2D version and the code is adapted to use the
3D visualization functions of empyre by setting z to 0.

Since empyre creates an ndarray from the input tuple, having unequal length
or incompatible shapes will trigger "Creating an ndarray from ragged nested
sequences (which is a list-or-tuple of lists-or-tuples-or ndarrays with
different lengths or shapes) is deprecated." since NumPy 1.19.

For that reason we use np.broadcast_arrays() to make the shapes and sizes match.

CC @jan-car :-)