Update azure pipelines (#908)
* can't show pip version before tox environment exists...

* use default pipeline trigger

* Try to enable caching for pip (should also cache built wheels)

This should speed up tox environment creation for subsequent builds.

* Try to re-add limiting ci trigger

* Try harder to cache things

* Re-use dask scheduler and workers across test session

This is a trade-off: the workers aren't as 'pristine' as when we start
a new set of workers for each test case, but on the other hand we can
share the numba compiled functions and other per-process caches. Let's
see how this works out.