Use the right loop in dd.as_completed
This is needed if there are multiple `Client`s running, as the `Client`
that was created last is registered as the global `Client` and then used
for functions like `dd.as_completed`. That means we create futures on
a different loop than we try to retrieve their results, which results in
hangs (can be debugged using PYTHONASYNCIODEBUG=1), as setting the
result for a future is not threadsafe.