UDF: Consistency fixes and `get_results` improvements
Now, all accesses to buffers in `UDF.results` and the `src`/`dest`
parameters of `UDF.merge` should be done via `__getattr__`; if the
`BufferWrapper` is really needed, then `UDF.results.get_buffer` should
be used.

As a follow-up to #994, `get_results` now should consistently return
numpy arrays, which are then converted to `BufferWrapper`s before
returning them to the user. This means we can make `get_results` a bit
more DRY by not having to repeat buffer names. Also, buffers that don't
have to be post-processed, can be left out in `get_results, and will be
returned unchanged. If you want to use a buffer only internally, without
exposing it to the user, pass `use='private'` to `UDF.buffer`.

The `allocate` parameter was merged with the `use` parameter, and now
`use='result_only'` should be used instead of `allocate=False`.

Fixes #1000, #1001, #631