WIP: Propagate damage through UDF execution
For setting vmin and vmax of nav buffers properly, the plot function
needs to know which data was already assigned. single and sig buffers
can be assumed to be completely damaged after each partition result has been merged.

Since the damage ist the same for all UDFs that are executed in the same pass, we only have to track
one kind='nav' damage buffer per pass.

This change contains infrastructure to pass the damage from UDFRunner to the plot.

The step is actually maintaining the damage information, and then using it.

Co-authored-by: Alexander Clausen <a.clausen@fz-juelich.de>