Improve fill value outside of ROI
Using np.nan for types other than floats led to unexpected results.
Now the values are hopefully more in line with what users would expect.

Technically this leads to changed results of calculations and could break user code.
However, the main use case of floating point values is unaffected and code that relies on the
previous fill values seems unlikely, overall.

To be discussed if we just do a hard change with 0.7.0 or go through depreciation.