Fix docstrings across LiberTEM
The `__init__()` method should be documented through the docstring of its class.
That way one doesn't need `:special-members: __init__`.
Sphinx started to throw
errors for automodules with that directive since it was looking for a top-level
`__init__()` function instead of `__init__()` methods in the classes of that module.

See also https://numpydoc.readthedocs.io/en/latest/format.html#class-docstring

Note that this is different from the original docstring
convention that calls for adding a docstring to `__init__()`:
https://www.python.org/dev/peps/pep-0257/