Resilient to float ROI
A ROI of type float triggered a cryptic Numba error deep within LiberTEM.
This can easily happen if a ROI is created with `np.zeros()` or `np.ones()`
without the `dtype=bool` argument.

With this change, a ROI that is not bool triggers a warning and is cast to
bool. This gives the behavior the user would expect for floats and ints:
Zero is False, non-zero True.