Limit the number of threads that are started
Using only threadpoolctl and similar methods to set thread limits
doesn't actually limit the number of threads that are started, but only
the number of threads that are actually used. This can become
problematic if you start many notebooks at once, each with their own
distributed cluster: it's easy to actually run into operating system
limits in cases like this.

By setting the `*_NUM_THREADS` environment variables before starting the
workers, we make sure to actually limit the number of threads that are
started.