Set NUMBA_NUM_THREADS in tests
That limits the threads in particular on machines with many cores
so that tests can run better in parallel. Furthermore it starts
Dask clusters with reduced number of threads per worker. Setting the
worker environment in a fixture for the cluster won't work since
Numba might be initialized already.

Make set_numba_threads robust against trying to set more threads than allowed.

This happens if NUMBA_NUM_THREADS is set and the InlineJobExecutor tries to
set the threads to the number of cores.