Disable upload of htmlcov for data tests
As the coverage is not merged upstream in azure pipelines, it is not so
useful - what would be useful is a complete combined coverage of all
jobs, merged by the python `coverage` tool itself, and generating
a usable `htmlcov` report from that. This could replace codecov.io very
well, but I don't think it is feasible at the moment.