Dynamic task scheduling
Tasks are assigned to workers at the time they are submitted. That means
a fast worker such as a GPU will get the same number of tasks assigned as a CPU
if all tasks are submitted at once.

Here, the ability of dask.distributed to add tasks dynamically to as_completed
is used to schedule more tasks only after previous tasks are completed. That way
they will be assigned to the worker that just became free so that faster workers
receive more work.

That yields a 20 % speed bonus in synthetic tests on an 18-core machine
with one GPU where a cuda worker completes a task 10x faster than a CPU worker.