Cache test dependencies
Since libertem/libertem:continuous changes all the time when the code
changes, we establish a base image only with the base requirements,
FIRST add the test requirements to it, and THEN install LiberTEM.
That allows to re-use the layer with the test requirements.

The production image is still tested for scheduler and workers.