Allow falling back to NumPy on all workers
Allow computing Dask arrays that are generated from UDF results with the DelayedJobExecutor
on a Distributed cluster that was set up with LiberTEM workers.

Background: The LiberTEM cluster setup uses resource tags to schedule tasks properly on both CPU and GPU workers
based on the capabilities of the UDF. If the appropriate resource requirements are not passed to
the `compute()` call, the scheduler will schedule the tasks on any workers without taking the UDF
capabilities into consideration.

With this change, UDFs that support NumPy (common case) will execute without specifying resource tags, i.e.
running them on CUDA workers will only trigger a warning.
For UDFs that don't support NumPy (i.e. only CuPy or CUDA), there's now a convenience function to generate
the resource tags.

Tests TODO after we settled on this or another solution.