Remove `compute()` for now
While working on examples and tests I didn't feel a need for this,
now that we have a fallback for CPU-only UDFs.
Also, relying on the most recent UDFs seems brittle and we should
maybe think of a better mechanism.

I'd propose to defer `compute()` for now and see how far we get without it.