TVIPS: Automatically detect `sync_offset` and `nav_shape` parameters
The detection works as follows:

`sync_offset` is the position where the scan coordinates go from
`(some_y, some_x)` to `(0, 0)` and then `(0, 1)`. Before the `(0, 0)`
position, there can be an arbitrary amount of values with undefined
values, which should be constant across images. Then, once we detect the
change from `(0, 0)` to `(0, 1)`, this is assumed to be the beginning of
the rectangular scan.

`nav_shape` is then built from the number of images in the file, minus
the `scan_offset`, and using the largest x-scan-coordinate as the width
of our shape.