Fix for validation messages not showing up
This was a multi-part issue - on one hand, the error messages from the
validation via JSON schema were not propagated properly to formik, on
the other hand, formik requires marking fields as "touched" - this was
not happening with `TupleInput` + `TupleInputPart`, which caused the
validation errors to be swallowed.

As an aside, I've updated the Formik library to a current version.