PipelinedExecutor: properly clean up response queue
As tracked down by @matbryan52 in https://github.com/LiberTEM/LiberTEM/pull/1267#issuecomment-1181828475
there was an issue where an exception while running a task would cause
the response queue to be polluted by error responses. This caused
subsequent `run_tasks` calls to fail immediately, as they drained the
error responses from the previous run. This is now fixed by properly
draining all responses for in-flight requests on any exception.

If there are similar issues in the future, we could switch to using
a UUID to identify the tasks, and ignoring (`logger.warning` maybe?) any
messages that belong to other tasks.