ApplyShiftedMasksUDF
Descan compensation for all applications that rely on ApplyMasksUDF

Utility method in libertem.common.slice to calculate a pair of slices
that select the intersection of two slices on both sides.

This is used in ApplyShiftedMasksUDF to select the overlap of frame and mask stack.
The API and functionality of ApplyShiftedMasksUDF is kept compatible with
ApplyMasksUDF as far as possible.