Avoid warning on casting complex to real
The dtype argument here is meant to force calculation with high precision even
if the input data is low precision. However, setting it to float64 will give wrong results
with complex input, and setting it to complex yields results with complex dtype
although the variance is always a non-negative real number
https://en.wikipedia.org/wiki/Complex_random_variable#Variance_and_pseudo-variance

To not trigger a warning about unsafe casting, this change drops the imaginary part explicitly.