Forbid arbitrary tiling changes when pixel correction is active
* Correct the dim check in validate
* Additional check after dataset.adjust_tileshape()

A reasonable dataset should only accept a tile shape as-is or set it to
full frames since it can't know if it may collide with correction
of bad pixels.

Many tests with memory dataset rely extensively on overriding tile shape
negotiation, so this only triggers an exception if corrections are
active. Otherwise it generates a warning.

There were some spurious issues with tests on my system that couldn't be
reproduced separately.