Keep previous behavior of backends parameter of run_udf
Previously, specifying backends only had effect on the UDF side
since the dataset would always generate NumPy which would be converted
to the required backend.

This change makes the API backwards-compatible. Datasets are free in
their choice of backends and the execution plan will use conversions as
necessary. Restrictions on the dataset
can still be imposed by dataset parameters, like for example the current
MemoryDataSet.