Prototype for dynamic parameter updates
This allows to dynamically update UDF parameters while LiberTEM is
running the UDF. Only implemented for the `PipelinedExecutor` for now.

With this patch applied, the following code works:

    async for partial_results in self.ctx.run_udf_iter(dataset=aq, udf=udfs, sync=False):
        # XXX hacks: parameter update
        udfs = list(self.udfs.values())
        if self.ctx._runner._params._kwargs != [udf._kwargs for udf in udfs]:
            key = self.ctx._runner._params_handle
            self.ctx._runner._params.update_from_udfs(udfs)
            self.ctx.executor.scatter_update(key, self.ctx._runner._params)

In combination with LiberTEM-live, this gives you interactivity while
a scan is running.

(this is not the final API or implementation approach, just a study that
this is possible with relatively moderate amount of changes)