Improved scheduling for the `PipelinedExecutor`
As discussed before, an easy improvements on the round-robin scheduling
implemented before, is sending tasks always to the process with the
shortest request queue. This is implemented in these changes.

In addition, this also speeds up the tests by a lot; on the system
I tested a full `tests_pipelined` run goes from ~108s to ~28s - by using
less worker processes, and by catching errors on startup faster. Also
marked the "default spec" case as slow, as that will always try to start
~many workers on large machines.