Typing drive-by fixes
`Tuple[T]` means exactly a one-tuple with a value of type `T`, which
is in most cases not what is actually meant by the annotations (compare
with `List[T]` which is a list of zero or more values of type `T`).

Replace instances where this is clearly the case with
`Tuple[T, ...]`, which means a tuple with zero or more values of
type `T`.

Also some minor drive-by typing fixes.