Fix _baked_revision building
We use `python -m build` these days, which uses PEP517 isolated builds.
This means the `build_py` command is executed not on the source tree,
but on a copy built from the sdist. There, we can't get the git revision
anymore...
It also turns out that the baked revision in the sdist was written to
a bad path (libertem/_baked_revision.py vs
src/libertem/_baked_revision.py), and by fixing that, the wheel now also
contains the correct revision.