Ensure liveness of executor for some utility endpoints
Setting the snooze timeout very low, for example by
starting with `libertem-server --snooze-timeout 0.001`, exposes some
issues which this commit fixes, namely that the
executor snoozes before `get_fs_listing`/`detect` finish running.

It's not recommended to set the snooze timeout so small, but this change
should increase stability in these cases - also possibly useful in cases
where a lot of time passes between getting the executor and when the
actual function is called, for example when the system is put to sleep, or
possibly in case of forward jumps in time (?)

Refs #1576 - lower-level keep-alive would prevent this kind of issue
completely.