Ensure we properly close generators/iterators
If we don't, we may swallow exceptions. This should improve reliability.

Without this, running `pytest tests/executor/test_pipelined.py -n12`
hung quite reproducibly in `CommsDispatcher.__exit__`, waiting for the
`monitor_queue` thread to return. I'm not 100% sure how this hang came
to be, but this appears to be fixed by proper exception handling.