Allow small partitions
In case a `DataSet` has less frames than partitions (i.e. workers),
we previously combined the frames into a single partition. This is
a performance issue if there is a lot of work per-frame. Instead, create
one partition per frame so we get more parallelization.

See #1701 for some initial discussion and other options.