Increase timeout for `test_libertem_server_cli_startup`
Usually (in the last 14? months) the process dies inside a one-second window,
but it now happened once that the deadline was missed, so increase the timeout
to 30 seconds. If the CI runner is slow, we can't do much, it's just important
that we eventually fail if the process doesn't die here.