Allow monitoring with AutoUDF
This is intended for live processing, in particular if the auto function
returns a 2D result, such as a filter applied to each frame.

Not sure if this is generally useful as-is, but I needed to give an example for Natalia
how to do live monitoring, and why not turn it into an actual feature.

I feel like somehow an update monitor could be a general feature of buffers.
For `kind="nav"` and `kind="single"` a natural update has `extra_shape`,
and for `kind="sig"` it is the buffer shape itself. This could actually
go somewhere in the merge and/or live plotting code.

However, maybe each udf should just implement the `monitor` parameter itself,
like here, and decide what actually makes sense to monitor. It doesn't seem very
useful for UDFs like `CoMUDF`, for example.

Putting this up for discussion here!

Drive-by:
* Fix log.warn() deprecation
* Remove arbitrary size warning. LiberTEM executors can handle large results
  more efficiently these days, in particular threaded and pipelined.