Delegate GPU selection to Executor and Environment
With this change, the UDFPartRunner just queries Environment
and doesn't interact with environment variables etc directly.

This opens the door to enable GPU processing on all executors, incl. threaded
and inline, in a proper way. For now they query the environment variables
to reproduce the previous behavior.

In a next step, I'll look into using an Environment also for the main process.