Change venv/TOX_WORK_DIR/git directories
Because the gitlab runner helpfully cleans up the working directory, as
it should, we need to put the venv, the TOX_WORK_DIR and the reference
git clone into different directories. The alternative would have been to
use artifacts, but I decided against this because

1) There are many files in these directories, and
2) They would also be transferred to the GitHub actions run by default

So instead, we use scratch space for this. If this works, we also need
an `after_script` to clean things up.